Zmieniono int na unsigned long w typie mapy

+++ strqueue.cpp
@@ -7,7 +7,7 @@
 
 using namespace std;
 
-static unordered_map<int, deque<string>> Map;
+static unordered_map<unsigned long, deque<string>> Map;
 
 unsigned long strqueue_new() {
     // TODO
